version 8.1 Refactor quantity and price handling across controllers and models to support decimal values with up to four decimal places. Update validation rules for quantity and price fields to ensure proper numeric input. Adjust CSV export formats to reflect these changes. Enhance data consistency in stock and transaction models by casting relevant fields to float.

+++ app/Models/StockBatch.php
@@ -22,9 +22,9 @@ class StockBatch extends Model
 
     protected $casts = [
         'purchase_date' => 'date',
-        'unit_price' => 'integer',
-        'qty_purchased' => 'integer',
-        'qty_available' => 'integer',
+        'unit_price' => 'float',
+        'qty_purchased' => 'float',
+        'qty_available' => 'float',
     ];
 
     public function purchaseLine(): BelongsTo
